ams-api/dapi/requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
dnspython==1.16.0
pymongo==3.10.1
pytz==2019.3
sqlparse==0.3.1
